+++ .env.local.example
@@ -1,5 +1,2 @@
  NEXT_PUBLIC_SUPABASE_URL=your-project-url
  NEXT_PUBLIC_SUPABASE_ANON_KEY=your-project-key
- 
-
- https://xnjgzwpzppkttqesxmhj.supabase.co
