+++ app/page.tsx
@@ -1,5 +1,3 @@
-import SignUpForm from "@components/signup-form";
-
 export default function Home() {
   return (
     <main>
