+++ app/routes/routes.tsx
@@ -1,0 +1,6 @@
+const routes = {
+  HOME: "/",
+  LOGIN: "/routes/login",
+};
+
+export default routes;
