+++ lib/components/navbar.tsx
@@ -1,6 +1,7 @@
 import { RxHamburgerMenu } from "react-icons/rx";
 import { PiCookingPotLight } from "react-icons/pi";
 import Link from "next/link";
+import routes from "app/routes/routes";
 
 export default function Navbar() {
   return (
@@ -34,7 +35,7 @@ export default function Navbar() {
               </a>
             </li>
             <li>
-              <Link href="/routes/login">
+              <Link href={routes.LOGIN}>
                 <button className="block text-white  hover:bg-green-700 p-2 rounded-xl bg-green-600 ">
                   Login
                 </button>
